- Added string with newlines rendering - Some refactoring

+++ Demo/src/l2hal/src/l2hal_errors.c
@@ -1,6 +1,6 @@
 #include <l2hal_errors.h>
 
-void L2HAL_Error(int8_t code)
+void L2HAL_Error(int16_t code)
 {
 	L2HAL_ERROR_CLOCK_SIGNAL_PORT();
